ref(network): move various helpers to peer book
