+++ network/src/peers/peers.rs
@@ -95,23 +95,6 @@ impl Node {
         Ok(())
     }
 
-    ///
-    /// Returns the `SocketAddr` of the last seen peer to be used as a sync node, or `None`.
-    ///
-    pub fn last_seen(&self) -> Option<SocketAddr> {
-        if let Some((&socket_address, _)) = self
-            .peer_book
-            .read()
-            .connected_peers()
-            .iter()
-            .max_by(|a, b| a.1.last_seen().cmp(&b.1.last_seen()))
-        {
-            Some(socket_address)
-        } else {
-            None
-        }
-    }
-
     async fn initiate_connection(&self, remote_address: SocketAddr) -> Result<(), NetworkError> {
         let own_address = self.local_address().unwrap(); // must be known by now
         if remote_address == own_address
@@ -237,7 +220,7 @@ impl Node {
         let current_block_height = self.consensus().current_block_height();
         let connected_peers = self.peer_book.read().connected_peers().clone();
         for (remote_address, _) in connected_peers {
-            self.sending_ping(remote_address);
+            self.peer_book.read().sending_ping(remote_address);
 
             self.outbound
                 .send_request(Message::new(
@@ -272,54 +255,6 @@ impl Node {
         }
     }
 
-    fn peer_quality(&self, addr: SocketAddr) -> Option<Arc<PeerQuality>> {
-        self.peer_book
-            .read()
-            .connected_peers()
-            .get(&addr)
-            .map(|peer| Arc::clone(&peer.quality))
-    }
-
-    fn sending_ping(&self, target: SocketAddr) {
-        if let Some(quality) = self.peer_quality(target) {
-            let timestamp = Instant::now();
-            *quality.last_ping_sent.lock() = Some(timestamp);
-            quality.expecting_pong.store(true, Ordering::SeqCst);
-        } else {
-            // shouldn't occur, but just in case
-            warn!("Tried to send a Ping to an unknown peer: {}!", target);
-        }
-    }
-
-    /// Handles an incoming `Pong` message.
-    pub fn received_pong(&self, source: SocketAddr) {
-        if let Some(quality) = self.peer_quality(source) {
-            if quality.expecting_pong.load(Ordering::SeqCst) {
-                let ping_sent = quality.last_ping_sent.lock().unwrap();
-                let rtt = ping_sent.elapsed().as_millis() as u64;
-                quality.rtt_ms.store(rtt, Ordering::SeqCst);
-                quality.expecting_pong.store(false, Ordering::SeqCst);
-            } else {
-                quality.failures.fetch_add(1, Ordering::Relaxed);
-            }
-        } else {
-            // shouldn't occur, but just in case
-            warn!("Received a Pong from an unknown peer: {}!", source);
-        }
-    }
-
-    ///
-    /// Updates the last seen timestamp of this peer to the current time.
-    ///
-    #[inline]
-    pub fn update_last_seen(&self, addr: SocketAddr) {
-        if let Some(ref quality) = self.peer_quality(addr) {
-            *quality.last_seen.write() = Some(chrono::Utc::now());
-        } else {
-            warn!("Attempted to update state of a peer that's not connected: {}", addr);
-        }
-    }
-
     /// TODO (howardwu): Implement manual serializers and deserializers to prevent forward breakage
     ///  when the PeerBook or PeerInfo struct fields change.
     ///
@@ -342,35 +277,6 @@ impl Node {
     //     Ok(())
     // }
 
-    /// Registers that the given number of blocks is expected as part of syncing with a peer.
-    pub fn expecting_sync_blocks(&self, addr: SocketAddr, count: usize) {
-        if let Some(ref pq) = self.peer_quality(addr) {
-            pq.remaining_sync_blocks.store(count as u16, Ordering::SeqCst);
-        } else {
-            error!("Peer for expecting_sync_blocks purposes not found!");
-        }
-    }
-
-    /// Registers the receipt of a sync block from a peer; returns `true` when finished syncing.
-    pub fn got_sync_block(&self, addr: SocketAddr) -> bool {
-        if let Some(ref pq) = self.peer_quality(addr) {
-            pq.remaining_sync_blocks.fetch_sub(1, Ordering::SeqCst) == 1
-        } else {
-            error!("Peer for got_sync_block purposes not found!");
-            true
-        }
-    }
-
-    /// Checks whether the current peer is involved in a block syncing process.
-    pub fn is_syncing_blocks(&self, addr: SocketAddr) -> bool {
-        if let Some(ref pq) = self.peer_quality(addr) {
-            pq.remaining_sync_blocks.load(Ordering::SeqCst) != 0
-        } else {
-            error!("Peer for got_sync_block purposes not found!");
-            false
-        }
-    }
-
     /// TODO (howardwu): Add logic to remove the active channels
     ///  and handshakes of the peer from this struct.
     /// Sets the given remote address in the peer book as disconnected from this node server.
@@ -379,7 +285,7 @@ impl Node {
     pub(crate) fn disconnect_from_peer(&self, remote_address: SocketAddr) -> Result<(), NetworkError> {
         debug!("Disconnecting from {}", remote_address);
 
-        if self.is_syncing_blocks(remote_address) {
+        if self.peer_book.read().is_syncing_blocks(remote_address) {
             self.consensus().finished_syncing_blocks();
         }
